% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmdb_response.R
\name{base_tmdb_request}
\alias{base_tmdb_request}
\title{base_tmdb_request}
\usage{
base_tmdb_request(auth_token = get_auth_token())
}
\arguments{
\item{auth_token}{}
}
\value{
httr2 request object with base URL, Authentication, and Throttling
parameters set
}
\description{
base_tmdb_request
}
