% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover_movies.R
\name{discover_movies}
\alias{discover_movies}
\title{Discover_movies}
\usage{
discover_movies(num_pages = 5, ...)
}
\arguments{
\item{num_pages}{Max. number of pages to request from API server
(default = 5 pages x 20 results/page)}

\item{...}{Search parameters as seen on documentation website}
}
\description{
\href{https://developer.themoviedb.org/reference/discover-movie}{Discover TMDB movies}
has a number of available optional parameters.
}
