% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmdb_response.R
\name{get_tmdb_response}
\alias{get_tmdb_response}
\title{get_tmdb_response}
\usage{
get_tmdb_response(path_args, ...)
}
\arguments{
\item{path_args}{Additional arguments appended to the base API url}

\item{...}{Other arguments passed as named query parameters}
}
\value{
The response in JSON format
}
\description{
get_tmdb_response
}
