% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_collection.R
\name{search_collection}
\alias{search_collection}
\title{search_collection}
\usage{
search_collection(query, ...)
}
\arguments{
\item{query}{Required search parameter}

\item{...}{Other optional search parameters}
}
\description{
\href{https://developer.themoviedb.org/reference/search-collection}{Search TMDB Collections database}
}
