% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movie_details.R
\name{movie_details}
\alias{movie_details}
\title{movie_details}
\usage{
movie_details(movie_id, append_to_response = NULL)
}
\arguments{
\item{movie_id}{Required search parameter, internal TMDB identifier for a movie}

\item{append_to_response}{\href{https://developer.themoviedb.org/docs/append-to-response}{Optional,} comma separated list of endpoints within this namespace, 20 items max (ex: \code{"videos,images"})}
}
\value{
JSON response for the movie details of specified movie
}
\description{
\href{https://developer.themoviedb.org/reference/movie-details}{Get TMDB movie details}
}
