% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_movies.R
\name{search_movies}
\alias{search_movies}
\title{search_movies}
\usage{
search_movies(query, ...)
}
\arguments{
\item{query}{Required search parameter}

\item{...}{Other optional search parameters}
}
\description{
\href{https://developer.themoviedb.org/reference/search-movie}{Search TMDB movie database}
Search for movies by their original, translated and alternative titles.
This function uses \code{\link[=search_movies_response]{search_movies_response()}} to get the data from one page of results,
then uses \code{\link[=call_tmdb_fn]{call_tmdb_fn()}} and \code{\link[=loop_through_pages]{loop_through_pages()}} to combine data from remaining
pages
}
\seealso{
\code{\link[=search_movies_response]{search_movies_response()}}, \code{\link[=get_tmdb_response]{get_tmdb_response()}}, \code{\link[=call_tmdb_fn]{call_tmdb_fn()}}, \code{\link[=unique]{unique()}}
}
